+++ src/bstone/src/bstone_sys_display_mode.cpp
@@ -1,0 +1,1 @@
+#include "bstone_sys_display_mode.h"