+++ src/bstone/src/bstone_sys_display_mode.h
@@ -1,0 +1,17 @@
+#ifndef BSTONE_SYS_DISPLAY_MODE_INCLUDED
+#define BSTONE_SYS_DISPLAY_MODE_INCLUDED
+
+namespace bstone {
+namespace sys {
+
+struct DisplayMode
+{
+	int width;
+	int height;
+	int refresh_rate;
+};
+
+} // namespace sys
+} // namespace bstone
+
+#endif // BSTONE_SYS_DISPLAY_MODE_INCLUDED