+++ src/bstone/src/bstone_sys_logger.cpp
@@ -11,18 +11,44 @@ SPDX-License-Identifier: MIT
 namespace bstone {
 namespace sys {
 
+namespace {
+
+constexpr const char* const empty_string = "";
+
+} // namespace
+
 Logger::Logger() = default;
 
 Logger::~Logger() = default;
 
 void Logger::log_information() noexcept
 {
-	log_information("");
+	log_information(empty_string);
 }
 
 void Logger::log_information(const char* message) noexcept
 {
 	do_log(LogLevel::information, message);
+}
+
+void Logger::log_warning() noexcept
+{
+	log_warning(empty_string);
+}
+
+void Logger::log_warning(const char* message) noexcept
+{
+	do_log(LogLevel::warning, message);
+}
+
+void Logger::log_error() noexcept
+{
+	log_error(empty_string);
+}
+
+void Logger::log_error(const char* message) noexcept
+{
+	do_log(LogLevel::error, message);
 }
 
 } // namespace sys